Remove some duplication.
